bot/source/requirements.txt
mysql-connector-python==9.0.0
python-telegram-bot==21.6
python-dotenv==1.0.1
loguru==0.7.2